Trying to fix some invalid left shift operations
Fixes issue #16

+++ src/imagew-bmp.c
@@ -425,7 +425,7 @@ static int find_high_bit(unsigned int x)
 {
 	int i;
 	for(i=31;i>=0;i--) {
-		if(x&(1<<i)) return i;
+		if(x&(1U<<(unsigned int)i)) return i;
 	}
 	return 0;
 }
@@ -433,7 +433,7 @@ static int find_low_bit(unsigned int x)
 {
 	int i;
 	for(i=0;i<=31;i++) {
-		if(x&(1<<i)) return i;
+		if(x&(1U<<(unsigned int)i)) return i;
 	}
 	return 0;
 }
